Python/pyproject.toml
[tool.poetry]
name = "gitlab"
version = "0.1.0"
description = ""
authors = ["Artur Akmalov <akmalov.mail@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
python-gitlab = "^3.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
